refactor(rollout): extract smooth-move helpers to shared utility
Move _teleop_supports_feedback, _teleop_smooth_move_to, and
_follower_smooth_move_to from dagger.py to lerobot/utils/smooth_move.py
so teleoperate, record, and replay can reuse them for soft handover.

Pure refactor: no behavior change, no new call sites.
